fixing warnings for unused variables

+++ StRoot/StFwdTrackMaker/XmlConfig/HistoBins.h
@@ -49,7 +49,7 @@ public:
    {
       if ( nullptr == _x ) return;
 
-      for ( unsigned i = 0; i <= _x->GetNbins(); i++ ) {
+      for ( unsigned int i = 0; i <= static_cast<unsigned int>(_x->GetNbins()); i++ ) {
          if ( i < _labels.size() )
             _x->SetBinLabel( i + 1, _labels[i].c_str() );
       }
